Добавлен функционал блокировок и задач, куча ошибок была исправлена

+++ app/src/main/java/com/example/kurstaskmanager/data/Converters.java
@@ -3,12 +3,11 @@ package com.example.kurstaskmanager.data;
 import androidx.room.TypeConverter;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 
 public class Converters {
     @TypeConverter
-    public static String fromList(List<Integer> list) {
+    public static String fromList(List<Long> list) {
         if (list == null || list.isEmpty()) return "";
         StringBuilder sb = new StringBuilder();
         for (int i = 0; i < list.size(); i++) {
@@ -19,12 +18,12 @@ public class Converters {
     }
 
     @TypeConverter
-    public static List<Integer> toList(String value) {
+    public static List<Long> toList(String value) {
         if (value == null || value.trim().isEmpty()) return new ArrayList<>();
-        List<Integer> list = new ArrayList<>();
+        List<Long> list = new ArrayList<>();
         for (String s : value.split(",")) {
             try {
-                list.add(Integer.parseInt(s.trim()));
+                list.add(Long.parseLong(s.trim()));
             } catch (NumberFormatException ignored) {}
         }
         return list;
